Got email working!

+++ Data/MyDbClasses.cs
@@ -33,7 +33,7 @@
         public string Username { get; set;}
         public string Password { get; set;}
         public string Hostname { get; set;}
-        public double Port { get; set; }
+        public int Port { get; set; }
         public string Recipient { get; set; }
     }
 }
